requirements.txt
PyWavelets
pydicom
pygame